Started working on PBW port to FGPU
I have added gpu:modifier element to the xslt spec.

In order to support backwards compatibility, the XSLT processor should be updated to a version which supports pulling in defaults from the schema (.xsd)

In this commit I have updated device side constants, PBM length and valid dimensions.

In theory, all that remains is to consider how best to iterate the bins.